refactor: LocationLikeCountService 리팩토링

+++ src/main/java/TourData/backend/domain/location/dto/LocationDto.java
@@ -8,7 +8,7 @@ public class LocationDto {
     public record LocationLikeCheckResponse (boolean isLike) {
     }
 
-    public record LocationLikeCountResponse (Integer likeCount) {
+    public record LocationLikeCountResponse (long likeCount) {
     }
 
 }
